+++ src/gql/resolvers/monthlyBalance.js
@@ -1,10 +1,8 @@
 'use strict';
 
-const { AuthenticationError } = require('apollo-server-express');
-
 const { logger } = require('../../utils/logger');
 
-const { Users, MonthlyBalance } = require('../../data/models/index');
+const { MonthlyBalance } = require('../../data/models/index');
 const { authValidations } = require('../auth/authValidations');
 
 /**
@@ -19,11 +17,7 @@ module.exports = {
 		getMonthlyBalance: async (root, args, context) => {
 			authValidations.ensureThatUserIsLogged(context);
 
-			const uuidOfUser = authValidations.getUserUUID(context);
-			const user = await Users.findOne({ uuid: uuidOfUser });
-			if (!user) {
-				throw new AuthenticationError('You must be logged in to perform this action');
-			}
+			const user = await authValidations.getUser(context);
 
 			try {
 				const allMonthlyBalance = await MonthlyBalance.find({ user_id: user._id }, null, { sort: { date: 1 } }).lean();
@@ -51,11 +45,7 @@ module.exports = {
 		registerMonthlyBalance: async (root, { balance, date }, context) => {
 			authValidations.ensureThatUserIsLogged(context);
 
-			const uuidOfUser = authValidations.getUserUUID(context);
-			const user = await Users.findOne({ uuid: uuidOfUser });
-			if (!user) {
-				throw new AuthenticationError('You must be logged in to perform this action');
-			}
+			const user = await authValidations.getUser(context);
 
 			return new MonthlyBalance({ user_id: user._id, balance, date }).save()
 				.then(monthlyBalance => {
@@ -77,13 +67,9 @@ module.exports = {
 		deleteMonthlyBalance: async (root, { uuid }, context) => {
 			authValidations.ensureThatUserIsLogged(context);
 
-			const uuidOfUser = authValidations.getUserUUID(context);
-			const user = await Users.findOne({ uuid: uuidOfUser });
-			if (!user) {
-				throw new AuthenticationError('You must be logged in to perform this action');
-			}
+			const user = await authValidations.getUser(context);
 
-			return MonthlyBalance.findOneAndDelete({ uuid })
+			return MonthlyBalance.findOneAndDelete({ uuid, user_id: user._id })
 				.then(monthlyBalance => {
 					return {
 						balance: monthlyBalance.balance.toString(),
@@ -103,11 +89,7 @@ module.exports = {
 		deleteAllMonthlyBalances: async (root, args, context) => {
 			authValidations.ensureThatUserIsLogged(context);
 
-			const uuidOfUser = authValidations.getUserUUID(context);
-			const user = await Users.findOne({ uuid: uuidOfUser });
-			if (!user) {
-				throw new AuthenticationError('You must be logged in to perform this action');
-			}
+			const user = await authValidations.getUser(context);
 
 			try {
 				return await MonthlyBalance.deleteMany({ user_id: user._id });
